+++ src/app/Configuracoes/page.tsx
@@ -1,0 +1,18 @@
+'use client';
+
+import { Suspense } from 'react';
+import Sidebar from '../components/Sidebar/sidebar';
+import ConfiguracoesContent from './ConfiguracoesContent';
+
+export default function Configuracoes() {
+  return (
+    <>
+      <div className="flex min-h-screen bg-[#E8E6E6]">
+        <Sidebar />
+        <Suspense fallback={<div className="flex-1 p-6">Carregando...</div>}>
+          <ConfiguracoesContent />
+        </Suspense>
+      </div>
+    </>
+  );
+}
